+++ CHANGELOG.md
@@ -12,6 +12,27 @@ The format of this changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/).
 
+## In progress
+
+#### New Features
+
+  - Improve hybrid nonlinear eigensolver convergence when using 2nd-order absorbing boundary
+    conditions. [PR 792](https://github.com/awslabs/palace/pull/792).
+
+#### Interface Changes
+
+  - Eigenmode simulations with frequency-dependent boundary conditions (e.g. waveports,
+    second-order absorbing, surface conductivity) now evaluate the boundary conditions at
+    the true complex frequency/eigenmode. Previously, the boundary conditions were
+    evaluated using only the real component of the eigenmode.
+    [PR 778](https://github.com/awslabs/palace/pull/778).
+
+#### Bug Fixes
+
+  - Fixed a bug in the 2D mode eigensolver that sometimes led to crashes when waveports
+    were used in conjunction with non-zero conductivity materials.
+    [PR 778](https://github.com/awslabs/palace/pull/778).
+
 ## [0.17.0] - 2026-06-28
 
 #### New Features